celery-requirements.txt
amqp==5.1.1
celery==5.3.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.5
django-celery-results==2.5.1
fasteners==0.14.1
redis==5.0.0